+++ CHANGELOG.md
@@ -8,9 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Features
 - Support `TestAdapter` custom executer.
 - Support environment variables to overwrite settings in the `TestAdapter`. (Fix: #57)
+### Command
+- Update to use `JsonService` implementation to use `Newtonsoft.Json` in the domain.
 ### TestAdapter
 - Use `LocalExtensionData` to store `TestModel` result.
 - Add `EnvironmentSettings` to overwrite settings in the `TestAdapter`.
+- Update `JsonService` implementation to use `Newtonsoft.Json` in the domain.
 
 ## [1.5.0] / 2024-09-11 - 2024-09-19
 ### Features